+++ ws/checkers/__init__.py
@@ -2,6 +2,7 @@
 
 from .CheckerBase import *
 from .ExtlinkStatusChecker import *
+from .ExtlinkStatusUpdater import *
 from .ExtlinkReplacements import *
 from .ManTemplateChecker import *
 from .WikilinkChecker import *